added control lose to movement and other pickables

+++ Source/AndroidRunner/ShieldBase.h
@@ -16,10 +16,10 @@ class ANDROIDRUNNER_API AShieldBase : public APickableBase
 	
 public: 
 	UPROPERTY(EditAnywhere, Category = "Basic")
-		int32 Armor = 100;
+		float Armor = 100;
 
 	UPROPERTY(EditAnywhere, Category = "Basic")
-		int32 Duration = 5.f;
+		float Duration = 5.f;
 
 	virtual void PawnOverlap(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult & SweepResult) override;
